se añadio al gitignore application.dev.properties ya que ahora contiene las variables de entorno, se creo la clase ApiResponse la cual crea la respuesta estandar a devolver en cada operacion, el caso de uso eliminar usuario ahora añade logica para cuando no exista el usuario lanzar una excepcion de dominio, se creo la clase DomainException la cual maneja todos los errores que pueden ocurrir a nivel de doinio, se creo la calse ErrorMessageCode la cual contiene los mensajes a mostrar segun el tipo de error, se añadio el mmanejo de errores a los casos de uso de buscar usuario , se creo una clase GlobalExceptionHandler la cual dispara los errores del sistema de forma automatica segun la case y se añadio metodos que detecta errores de Dto aso como del dominio deu usuario y en sentido general del sistema,en el jpa del usuari se añadio la verificacion de si el usuario existe por id o por email antes de crearlo , asi como para actualizarlo se verifica cuales campos fueron enviados y en caso de todo estar correcto actualizao y para eliminar verifica que el usuario existe o no por id y sin no lanza un error ,  se creo un metodo ResponseHandler el cual englobal el apiResponse para responer de forma personalizada los mensajes, el updateUserDto fue actualizado para tener todos los campos opcionales con sus seters y getes asi como las validaciones faltantes, se añadio el metodo toString para poder observar los objetos tanto en User como en UserEntity, se actualizo el userReposittoryPort para que debuelva boolean si encuentra o no un usuario con ese id existente , se actualizo el mapper de un dto a User de domain para ognorar los capons faltantes en un updateDto, se elimino el userNotFoundException para evitar redundancias

+++ src/main/java/com/firstSpringBootProject/first/User/infrastructure/adapters/output/persistence/repositories/JpaUserRepositoryAdapter.java
@@ -1,5 +1,7 @@
 package com.firstSpringBootProject.first.User.infrastructure.adapters.output.persistence.repositories;
 
+import com.firstSpringBootProject.first.User.domain.exceptions.DomainException;
+import com.firstSpringBootProject.first.User.domain.exceptions.ErrorMessageCode;
 import com.firstSpringBootProject.first.User.domain.models.User;
 import com.firstSpringBootProject.first.User.domain.ports.out.UserRepositoryPort;
 import com.firstSpringBootProject.first.User.infrastructure.adapters.output.persistence.entities.UserEntity;
@@ -22,8 +24,35 @@ public class JpaUserRepositoryAdapter implements UserRepositoryPort {
 
     @Override
     public User save(User user) {
-        return userMapper.toUser(userPersistence.save(userMapper.toUserEntity(user)));
+        if (this.findByEmail(user.getEmail()).isPresent()) {
+
+            String message = " User with email " + user.getEmail() + " already exist";
+
+            throw new DomainException(
+                    ErrorMessageCode.EMAIL_CONFLICT.getStatus(),
+                    ErrorMessageCode.PERSISTENCE_ERROR.getType(),
+                    message
+                    , null
+            );
+
+        }
+
+        if (this.findByUsername(user.getUsername()).isPresent()) {
+
+            String message = " User with username " + user.getUsername() + " already exist";
+
+            throw new DomainException(
+                    ErrorMessageCode.USERNAME_CONFLICT.getStatus(),
+                    ErrorMessageCode.USERNAME_CONFLICT.getType(),
+                    message
+                    , null
+            );
+        }
+
+        UserEntity entity = userPersistence.save(userMapper.toUserEntity(user));
+        return userMapper.toUser(entity);
     }
+
 
     @Override
     public List<User> findAll() {
@@ -34,7 +63,6 @@ public class JpaUserRepositoryAdapter implements UserRepositoryPort {
     public Optional<User> findById(Long id) {
         return userPersistence.findById(id)
                 .map(userMapper::toUser);
-//        return Optional.ofNullable(userMapper.toUser(userPersistence.findById(id).get()));
     }
 
     @Override
@@ -51,25 +79,58 @@ public class JpaUserRepositoryAdapter implements UserRepositoryPort {
 
     @Override
     public User update(Long id, User user) {
+
+        String message = "Updating User with ID " + id + " was not found";
+
         UserEntity existingEntity = userPersistence.findById(id)
-                .orElseThrow(() -> new RuntimeException("Updating User with ID " + id + " was not found"));
 
-        existingEntity.setEmail(user.getEmail());
-        existingEntity.setUsername(user.getUsername());
-        existingEntity.setFirstName(user.getFirstName());
-        existingEntity.setLastName(user.getLastName());
-        existingEntity.setAddress(user.getAddress());
-        existingEntity.setCellphone(user.getCellphone());
-        existingEntity.setPassword(user.getPassword());
-        existingEntity.setUserType(user.getUserType());
+                .orElseThrow(() -> new DomainException(
+                                        ErrorMessageCode.USER_NOT_FOUND.getStatus(),
+                                        ErrorMessageCode.USER_NOT_FOUND.getType(),
+                                        message,
+                                        null
+                                ));
 
-        UserEntity updatedEntity = userPersistence.save(existingEntity);
+        if(user.getEmail() != null){
+            existingEntity.setEmail(user.getEmail());
+        }
+        if(user.getUsername() != null){
+            existingEntity.setUsername(user.getUsername());
+        }
+        if(user.getFirstName() != null){
+            existingEntity.setFirstName(user.getFirstName());
+        }
+        if(user.getLastName() != null){
+            existingEntity.setLastName(user.getLastName());
+        }
+        if(user.getAddress() != null){
+            existingEntity.setAddress(user.getAddress());
+        }
+        if(user.getCellphone() != null){
+            existingEntity.setCellphone(user.getCellphone());
+        }
+        if(user.getPassword() != null){
+            existingEntity.setPassword(user.getPassword());
+        }
+        if(user.getUserType() != null){
+            existingEntity.setUserType(user.getUserType());
+        }
+
+         UserEntity updatedEntity = userPersistence.save(existingEntity);
         return userMapper.toUser(updatedEntity);
-       /* return null;*/
     }
 
     @Override
-    public void deleteById(Long id) {
-      userPersistence.deleteById(id);
+    public boolean deleteById(Long id) {
+     // userPersistence.deleteById(id);
+     // return false;
+        Optional<User> existingUser = this.findById(id);
+
+        if (existingUser.isPresent()) {
+            userPersistence.deleteById(id); // Llama al método void de Spring Data JPA
+            return true; // El usuario existía y se intentó eliminar
+        } else {
+            return false; // El usuario no fue encontrado
+        }
     }
 }
